split HomeSongs into songs on device and the rest
