+++ composeApp/src/androidMain/kotlin/app/kreate/android/themed/rimusic/screen/home/HomeSongs.kt
@@ -1,0 +1,324 @@
+package app.kreate.android.themed.rimusic.screen.home
+
+import androidx.compose.foundation.ExperimentalFoundationApi
+import androidx.compose.foundation.background
+import androidx.compose.foundation.layout.PaddingValues
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.lazy.LazyColumn
+import androidx.compose.foundation.lazy.LazyListState
+import androidx.compose.foundation.lazy.itemsIndexed
+import androidx.compose.foundation.text.BasicText
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.LaunchedEffect
+import androidx.compose.runtime.derivedStateOf
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.mutableStateOf
+import androidx.compose.runtime.remember
+import androidx.compose.runtime.saveable.rememberSaveable
+import androidx.compose.runtime.setValue
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.Brush
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.platform.LocalContext
+import androidx.compose.ui.text.style.TextOverflow
+import androidx.compose.ui.unit.dp
+import androidx.compose.ui.util.fastFilter
+import androidx.compose.ui.util.fastMap
+import androidx.media3.common.util.UnstableApi
+import androidx.media3.exoplayer.offline.Download
+import androidx.navigation.NavController
+import it.fast4x.compose.persist.persistList
+import it.fast4x.rimusic.Database
+import it.fast4x.rimusic.EXPLICIT_PREFIX
+import it.fast4x.rimusic.LocalPlayerServiceBinder
+import it.fast4x.rimusic.colorPalette
+import it.fast4x.rimusic.enums.BuiltInPlaylist
+import it.fast4x.rimusic.enums.DurationInMinutes
+import it.fast4x.rimusic.enums.MaxTopPlaylistItems
+import it.fast4x.rimusic.enums.SongSortBy
+import it.fast4x.rimusic.models.Song
+import it.fast4x.rimusic.service.MyDownloadHelper
+import it.fast4x.rimusic.service.modern.LOCAL_KEY_PREFIX
+import it.fast4x.rimusic.service.modern.isLocal
+import it.fast4x.rimusic.thumbnailShape
+import it.fast4x.rimusic.typography
+import it.fast4x.rimusic.ui.components.SwipeablePlaylistItem
+import it.fast4x.rimusic.ui.components.tab.toolbar.Button
+import it.fast4x.rimusic.ui.items.SongItemPlaceholder
+import it.fast4x.rimusic.ui.styling.Dimensions
+import it.fast4x.rimusic.ui.styling.onOverlay
+import it.fast4x.rimusic.ui.styling.overlay
+import it.fast4x.rimusic.utils.MaxTopPlaylistItemsKey
+import it.fast4x.rimusic.utils.Preference
+import it.fast4x.rimusic.utils.Preference.HOME_SONGS_SORT_BY
+import it.fast4x.rimusic.utils.Preference.HOME_SONGS_SORT_ORDER
+import it.fast4x.rimusic.utils.addNext
+import it.fast4x.rimusic.utils.asMediaItem
+import it.fast4x.rimusic.utils.center
+import it.fast4x.rimusic.utils.color
+import it.fast4x.rimusic.utils.durationTextToMillis
+import it.fast4x.rimusic.utils.enqueue
+import it.fast4x.rimusic.utils.excludeSongsWithDurationLimitKey
+import it.fast4x.rimusic.utils.forcePlayAtIndex
+import it.fast4x.rimusic.utils.includeLocalSongsKey
+import it.fast4x.rimusic.utils.isDownloadedSong
+import it.fast4x.rimusic.utils.manageDownload
+import it.fast4x.rimusic.utils.parentalControlEnabledKey
+import it.fast4x.rimusic.utils.rememberPreference
+import it.fast4x.rimusic.utils.semiBold
+import kotlinx.coroutines.Dispatchers
+import kotlinx.coroutines.flow.distinctUntilChanged
+import kotlinx.coroutines.flow.flowOf
+import kotlinx.coroutines.flow.flowOn
+import kotlinx.coroutines.flow.map
+import kotlinx.coroutines.flow.onEach
+import me.knighthat.component.SongItem
+import me.knighthat.component.Sort
+import me.knighthat.component.song.PeriodSelector
+import me.knighthat.component.tab.DeleteAllDownloadedSongsDialog
+import me.knighthat.component.tab.DownloadAllSongsDialog
+import me.knighthat.component.tab.ExportSongsToCSVDialog
+import me.knighthat.component.tab.HiddenSongs
+import me.knighthat.component.tab.ItemSelector
+import me.knighthat.component.tab.Search
+import me.knighthat.database.ext.FormatWithSong
+
+@UnstableApi
+@ExperimentalFoundationApi
+@Composable
+fun HomeSongs(
+    navController: NavController,
+    builtInPlaylist: BuiltInPlaylist,
+    lazyListState: LazyListState,
+    itemSelector: ItemSelector<Song>,
+    search: Search,
+    buttons: MutableList<Button>,
+    itemsOnDisplay: MutableList<Song>,
+    getSongs: () -> List<Song>,
+) {
+    // Essentials
+    val binder = LocalPlayerServiceBinder.current
+    val context = LocalContext.current
+
+    //<editor-fold defaultstate="collapsed" desc="Settings">
+    val parentalControlEnabled by rememberPreference( parentalControlEnabledKey, false )
+    val maxTopPlaylistItems by rememberPreference( MaxTopPlaylistItemsKey, MaxTopPlaylistItems.`10` )
+    val includeLocalSongs by rememberPreference( includeLocalSongsKey, true )
+    val excludeSongWithDurationLimit by rememberPreference( excludeSongsWithDurationLimitKey, DurationInMinutes.Disabled )
+    //</editor-fold>
+
+    var items by persistList<Song>( "home/songs" )
+
+    val songSort = Sort ( HOME_SONGS_SORT_BY, HOME_SONGS_SORT_ORDER )
+    val topPlaylists = PeriodSelector( Preference.HOME_SONGS_TOP_PLAYLIST_PERIOD )
+    val hiddenSongs = HiddenSongs()
+    val exportDialog = ExportSongsToCSVDialog(
+        playlistName = builtInPlaylist.text,
+        songs = getSongs
+    )
+    val downloadAllDialog = DownloadAllSongsDialog( getSongs )
+    val deleteDownloadsDialog = DeleteAllDownloadedSongsDialog( getSongs )
+
+    /**
+     * This variable tells [LazyColumn] to render [SongItemPlaceholder]
+     * instead of [SongItem] queried from the database.
+     *
+     * This indication also tells user that songs are being loaded
+     * and not it's definitely not freezing up.
+     *
+     * > This variable should **_NOT_** be set to `false` while inside **first** phrase,
+     * and should **_NOT_** be set to `true` while in **second** phrase.
+     */
+    var isLoading by rememberSaveable { mutableStateOf(false) }
+
+    // This phrase loads all songs across types into [items]
+    // No filtration applied to this stage, only sort
+    LaunchedEffect( builtInPlaylist, topPlaylists.period, songSort.sortBy, songSort.sortOrder, hiddenSongs.isFirstIcon ) {
+        isLoading = true
+
+        val retrievedSongs = when( builtInPlaylist ) {
+            BuiltInPlaylist.All -> Database.songTable
+                                           .sortAll( songSort.sortBy, songSort.sortOrder, excludeHidden = hiddenSongs.isHiddenExcluded() )
+                                           .map { list ->
+                                               // Include local songs if enabled
+                                               list.fastFilter {
+                                                   !includeLocalSongs || !it.id.startsWith( LOCAL_KEY_PREFIX, true )
+                                               }
+                                           }
+
+            BuiltInPlaylist.Downloaded -> {
+                // [MyDownloadHelper] provide a list of downloaded songs, which is faster to retrieve
+                // than using `Cache.isCached()` call
+                val downloaded: List<String> = MyDownloadHelper.downloads
+                                                               .value
+                                                               .values
+                                                               .filter { it.state == Download.STATE_COMPLETED }
+                                                               .fastMap { it.request.id }
+                Database.songTable
+                        .sortAll( songSort.sortBy, songSort.sortOrder )
+                        .map { list ->
+                            list.fastFilter { it.id in downloaded }
+                        }
+            }
+
+            BuiltInPlaylist.Offline -> Database.formatTable
+                                               .sortAllWithSongs( songSort.sortBy, songSort.sortOrder, excludeHidden = hiddenSongs.isHiddenExcluded() )
+                                               .map { list ->
+                                                   list.fastFilter {
+                                                       val contentLength = it.format.contentLength ?: return@fastFilter false
+                                                       binder?.cache?.isCached( it.song.id, 0, contentLength ) == true
+                                                   }.map( FormatWithSong::song )
+                                               }
+
+            BuiltInPlaylist.Favorites -> Database.songTable.sortFavorites( songSort.sortBy, songSort.sortOrder )
+
+            BuiltInPlaylist.Top -> Database.eventTable
+                                           .findSongsMostPlayedBetween(
+                                               from = topPlaylists.period.timeStampInMillis(),
+                                               limit = maxTopPlaylistItems.toInt()
+                                           )
+                                           .map { list ->
+                                               // Exclude songs with duration higher than what [excludeSongWithDurationLimit] is
+                                               list.fastFilter { song ->
+                                                   excludeSongWithDurationLimit == DurationInMinutes.Disabled
+                                                           || song.durationText
+                                                                  ?.let { durationTextToMillis(it) < excludeSongWithDurationLimit.asMillis } == true
+                                               }
+                                           }
+
+            BuiltInPlaylist.OnDevice -> flowOf( emptyList() )
+        }
+
+        retrievedSongs.flowOn( Dispatchers.IO )
+                      .distinctUntilChanged()
+                      // Scroll list to top to prevent weird artifacts
+                      .onEach { lazyListState.scrollToItem( 0, 0 ) }
+                      .collect { items = it }
+    }
+
+    LaunchedEffect( items, search.inputValue ) {
+    items.filter { !parentalControlEnabled || !it.title.startsWith( EXPLICIT_PREFIX, true ) }
+         .filter {
+             // Without cleaning, user can search explicit songs with "e:"
+             // I kinda want this to be a feature, but it seems unnecessary
+             val containsTitle = it.cleanTitle().contains( search.inputValue, true )
+             val containsArtist = it.cleanArtistsText().contains( search.inputValue, true )
+
+             containsTitle || containsArtist
+         }
+        .let {
+            itemsOnDisplay.clear()
+            itemsOnDisplay.addAll( it )
+
+            isLoading = false
+        }
+    }
+
+    LaunchedEffect( builtInPlaylist ) {
+        val firstButton = if( builtInPlaylist == BuiltInPlaylist.Top ) topPlaylists else songSort
+        buttons.add( 0, firstButton )
+        buttons.add( 3, downloadAllDialog )
+        buttons.add( 4, deleteDownloadsDialog )
+        buttons.add( exportDialog )
+    }
+
+    //<editor-fold defaultstate="collapsed" desc="Dialog Renders">
+    exportDialog.Render()
+    downloadAllDialog.Render()
+    deleteDownloadsDialog.Render()
+    //</editor-fold>
+
+    LazyColumn(
+        state = lazyListState,
+        userScrollEnabled = !isLoading,
+        contentPadding = PaddingValues( bottom = Dimensions.bottomSpacer )
+    ) {
+        if( isLoading )
+            items(
+                count = 20,
+                key = { it }
+            ) { SongItemPlaceholder() }
+
+        itemsIndexed(
+            items = itemsOnDisplay,
+            key = { _, song -> song.id }
+        ) { index, song ->
+            val mediaItem = song.asMediaItem
+
+            val isLocal by remember { derivedStateOf { mediaItem.isLocal } }
+            val isDownloaded = isLocal || isDownloadedSong( mediaItem.mediaId )
+
+            SwipeablePlaylistItem(
+                mediaItem = mediaItem,
+                onPlayNext = { binder?.player?.addNext( mediaItem ) },
+                onDownload = {
+                    if( builtInPlaylist != BuiltInPlaylist.OnDevice ) {
+                        binder?.cache?.removeResource(mediaItem.mediaId)
+                        Database.asyncTransaction {
+                            formatTable.updateContentLengthOf( mediaItem.mediaId )
+                        }
+                        if ( !isLocal )
+                            manageDownload(
+                                context = context,
+                                mediaItem = mediaItem,
+                                downloadState = isDownloaded
+                            )
+                    }
+                },
+                onEnqueue = {
+                    binder?.player?.enqueue(mediaItem)
+                }
+            ) {
+                SongItem(
+                    song = song,
+                    itemSelector = itemSelector,
+                    navController = navController,
+                    modifier = Modifier.animateItem(),
+                    thumbnailOverlay = {
+                        if ( songSort.sortBy == SongSortBy.PlayTime || builtInPlaylist == BuiltInPlaylist.Top ) {
+                            var text = song.formattedTotalPlayTime
+                            var typography = typography().xxs
+                            var alignment = Alignment.BottomCenter
+
+                            if( builtInPlaylist == BuiltInPlaylist.Top ) {
+                                text = (index + 1).toString()
+                                typography = typography().m
+                                alignment = Alignment.Center
+                            }
+
+                            BasicText(
+                                text = text,
+                                style = typography.semiBold.center.color(colorPalette().onOverlay),
+                                maxLines = 2,
+                                overflow = TextOverflow.Ellipsis,
+                                modifier = Modifier
+                                    .fillMaxWidth()
+                                    .padding(horizontal = 8.dp, vertical = 4.dp)
+                                    .align(alignment)
+                                    .background(
+                                        brush = Brush.verticalGradient(
+                                            colors = listOf(
+                                                Color.Transparent,
+                                                colorPalette().overlay
+                                            )
+                                        ),
+                                        shape = thumbnailShape()
+                                    )
+                            )
+                        }
+                    },
+                    onClick = {
+                        search.hideIfEmpty()
+
+                        binder?.stopRadio()
+
+                        val mediaItems = getSongs().fastMap( Song::asMediaItem )
+                        binder?.player?.forcePlayAtIndex( mediaItems, index )
+                    }
+                )
+            }
+        }
+    }
+}
